+++ R/base-rev.R
@@ -15,19 +15,11 @@
 
 ################################################################################
 # FUNCTION:                 DESCRIPTION:
-#  rev,timeSeries            Reverts a 'timeSeries' object in time
+#  rev.timeSeries            Reverts a 'timeSeries' object in time
 ################################################################################
 
 
 rev.timeSeries <-  function(x) x[NROW(x):1,]
 
-
-setMethod("rev", "timeSeries", function(x) rev.timeSeries(x))
-
-
-## # until UseMethod dispatches S4 methods in 'base' functions
-## rev.timeSeries <- function(x) .rev.timeSeries(x)
-
-
-################################################################################
+setMethod("rev", "timeSeries", rev.timeSeries)